Cast prevLastEventOffset to ulong before comparison
This needs to be an unsigned integer before we can perform the
comparison, in order to satisfy a compiler warning about signedness.

+++ xml.cpp
@@ -353,7 +353,7 @@ public:
 
 	HRESULT UndoEventAdd()
 	{
-		if( lastEventOffset == 0 || prevLastEventOffset == ULONG_MAX ) {
+		if( lastEventOffset == 0 || (unsigned long)prevLastEventOffset == ULONG_MAX ) {
 
 			// Can't undo more than the very last event (no undo history).
 			return E_OUTOFMEMORY;
